GitHub Service II
Added a run history table to the ERD.

Added a new app setting for workflows.

Created the GitHub converter to convert labels to issue types.

Altered the name of the repository property on the Workflow Run Model.

Updated the data type of the Id property in the Workflow Run Model and Release Model.

Finished the GetIssues method.

Created the GetCommits method.

Created the GetPullRequests method.

Created the GetWorkflowRuns method.

Created the GetReleases method.

+++ GitHub Scraper/GitHubScraper/Models/Release Model.cs
@@ -6,7 +6,7 @@ namespace GitHubScraper.Models
     internal class ReleaseModel
     {
         public string? Repository { get; set; }
-        public required int Id { get; set; }
+        public required long Id { get; set; }
         public required string Name { get; set; }
         public required UserModel Author { get; set; }
         public required string Body { get; set; }
